refactor(daemon): isolate DB foundation lifecycle
