+++ platform/daemon/src/db-vacuum-worker.ts
@@ -1,0 +1,165 @@
+/**
+ * Post-readiness vacuum orchestration.
+ *
+ * This module may depend on the DB-owner runtime. Foundational vacuum state
+ * and SQLite primitives live in db-vacuum.ts so the accessor never imports
+ * this higher-level orchestration layer.
+ */
+
+import type { DbAccessor } from "./db-accessor";
+import type { DbOwnerClient } from "./db-owner-client";
+import { ownerQueryOne, ownerRunStatement, ownerTransaction } from "./db-owner-maintenance";
+import { dbOwnerIncrementalVacuum, dbOwnerVacuumConversion } from "./db-owner-runtime";
+import {
+	getVacuumConversionStatusAsync,
+	markVacuumConversionCompleted,
+	markVacuumConversionFailed,
+	markVacuumConversionRunning,
+	type VacuumConversionStatus,
+} from "./db-vacuum";
+import { logger } from "./logger";
+
+export interface IncrementalReclaimOptions {
+	readonly owner?: DbOwnerClient;
+	readonly batchPages?: number;
+	readonly maxBatches?: number;
+	readonly checkpointKey?: string;
+	readonly onCheckpoint?: (reclaimed: number, remaining: number) => void;
+}
+
+export interface VacuumConversionHandle {
+	readonly running: boolean;
+	stop(): void;
+	run(): Promise<VacuumConversionStatus>;
+}
+
+/** Reclaim free pages in bounded batches with a real durable resume record. Conversion remains monolithic. */
+export async function reclaimIncrementalVacuum(
+	accessor: DbAccessor,
+	opts: IncrementalReclaimOptions = {},
+): Promise<{ readonly reclaimed: number; readonly remaining: number }> {
+	const batchPages = Math.max(1, Math.min(10_000, Math.trunc(opts.batchPages ?? 1_000)));
+	const maxBatches = Math.max(1, Math.min(100_000, Math.trunc(opts.maxBatches ?? 100_000)));
+	const key = opts.checkpointKey ?? "vacuum.incremental-reclaim";
+	let reclaimed = 0;
+	let remaining = 0;
+	if (opts.owner) {
+		await ownerTransaction(opts.owner, "maintenance.vacuum.checkpoint.ensure", [
+			ownerRunStatement(
+				"CREATE TABLE IF NOT EXISTS db_vacuum_reclaim_checkpoints (checkpoint_key TEXT PRIMARY KEY, reclaimed INTEGER NOT NULL, remaining INTEGER NOT NULL, updated_at TEXT NOT NULL)",
+			),
+		]);
+		const saved = await ownerQueryOne<{ reclaimed: number; remaining: number }>(
+			opts.owner,
+			"maintenance.vacuum.checkpoint.read",
+			"SELECT reclaimed, remaining FROM db_vacuum_reclaim_checkpoints WHERE checkpoint_key = ?",
+			[key],
+		);
+		reclaimed = saved?.reclaimed ?? 0;
+		remaining = saved?.remaining ?? 0;
+	}
+	let previousRemaining: number | null = remaining > 0 ? remaining : null;
+	for (let batch = 0; batch < maxBatches; batch += 1) {
+		const before = previousRemaining;
+		if (opts.owner) remaining = await dbOwnerIncrementalVacuum(opts.owner, batchPages);
+		else {
+			if (!accessor.incrementalVacuumAsync) throw new Error("incremental vacuum operation is unavailable");
+			remaining = await accessor.incrementalVacuumAsync({ siteToken: "db-vacuum-worker.ts:incremental" });
+		}
+		const progressed = before === null ? Math.max(0, batchPages) : Math.max(0, before - remaining);
+		reclaimed += progressed;
+		previousRemaining = remaining;
+		if (opts.owner)
+			await ownerTransaction(opts.owner, "maintenance.vacuum.checkpoint.write", [
+				ownerRunStatement(
+					"INSERT INTO db_vacuum_reclaim_checkpoints (checkpoint_key, reclaimed, remaining, updated_at) VALUES (?, ?, ?, datetime('now')) ON CONFLICT(checkpoint_key) DO UPDATE SET reclaimed=excluded.reclaimed, remaining=excluded.remaining, updated_at=excluded.updated_at",
+					[key, reclaimed, remaining],
+				),
+			]);
+		opts.onCheckpoint?.(reclaimed, remaining);
+		if (remaining <= 0 || progressed <= 0) break;
+		await new Promise<void>((resolve) => setTimeout(resolve, 0));
+	}
+	return { reclaimed, remaining };
+}
+
+/**
+ * Start the post-ready one-shot conversion worker. The timer yields once so
+ * the listening callback can return and readiness can be recorded before work
+ * begins.
+ */
+export function startVacuumConversionWorker(
+	accessor: DbAccessor,
+	opts: { readonly startImmediately?: boolean; readonly owner?: DbOwnerClient } = {},
+): VacuumConversionHandle {
+	let active = true;
+	let timer: ReturnType<typeof setTimeout> | null = null;
+	let inFlight: Promise<VacuumConversionStatus> | null = null;
+
+	async function run(): Promise<VacuumConversionStatus> {
+		if (inFlight) return inFlight;
+		const cycle = (async (): Promise<VacuumConversionStatus> => {
+			const before = await getVacuumConversionStatusAsync(accessor);
+			if (before.state !== "pending" || before.attempts >= before.maxAttempts) return before;
+
+			await markVacuumConversionRunning(accessor);
+			const running = await getVacuumConversionStatusAsync(accessor);
+			if (running.state !== "running") return running;
+			logger.info("db-vacuum", "Post-ready conversion worker started", {
+				attempt: running.attempts,
+				maxAttempts: running.maxAttempts,
+			});
+
+			try {
+				if (opts.owner !== undefined) {
+					await dbOwnerVacuumConversion(opts.owner);
+				} else {
+					if (!accessor.vacuumConversionAsync) throw new Error("VACUUM conversion operation is unavailable");
+					await accessor.vacuumConversionAsync({ siteToken: "db-vacuum-worker.ts:conversion" });
+				}
+				await markVacuumConversionCompleted(accessor);
+				logger.info("db-vacuum", "Post-ready conversion worker completed");
+			} catch (error) {
+				const message = error instanceof Error ? error.message : String(error);
+				await markVacuumConversionFailed(accessor, message);
+				logger.error(
+					"db-vacuum",
+					"Post-ready conversion worker failed; retry is available on a later startup while the attempt budget remains",
+					error instanceof Error ? error : undefined,
+				);
+			}
+			return await getVacuumConversionStatusAsync(accessor);
+		})();
+		inFlight = cycle;
+		void cycle.then(
+			() => {
+				if (inFlight === cycle) inFlight = null;
+			},
+			() => {
+				if (inFlight === cycle) inFlight = null;
+			},
+		);
+		return cycle;
+	}
+
+	if (opts.startImmediately !== false) {
+		timer = setTimeout(() => {
+			if (!active) return;
+			void run().catch((error) => {
+				logger.error("db-vacuum", "Post-ready conversion worker crashed", error instanceof Error ? error : undefined);
+			});
+		}, 0);
+	}
+
+	return {
+		get running(): boolean {
+			return active;
+		},
+		stop(): void {
+			active = false;
+			if (timer) clearTimeout(timer);
+			timer = null;
+		},
+		run,
+	};
+}
